Move vh to viewholders package
